Added Shorthand Functions and permissions
