+++ src/events/client/threadUpdate.js
@@ -6,7 +6,6 @@ export default {
     name: Events.ThreadUpdate,
     once: false,
     execute: async (oldThread, newThread) => {
-        // If the thread is not created by the bot do nothing
         if (newThread.ownerId !== newThread.client.user.id) return;
 
         if (oldThread.archived && newThread.locked) {
@@ -55,14 +54,14 @@ export default {
 
         if (oldThread.locked && !newThread.locked) {
             await newThread.send({
-                content: `<:key:1098978684523778098> **${executor.username}** __unlocked__ this ${formattedTime}`,
+                content: `<:key:1098978684523778098> **${executor.username}** __unlocked__ this conversation ${formattedTime}`,
             });
         } else if (!oldThread.locked && newThread.locked) {
             if (executor.id === newThread.client.user.id) return;
             if (oldThread.archived && !newThread.archived) return;
 
             await newThread.send({
-                content: `<:lock:1098978659890626671> **${executor.username}** __locked__ this ${formattedTime}`,
+                content: `<:lock:1098978659890626671> **${executor.username}** __locked__ this conversation ${formattedTime}`,
             });
         }
     },
